+++ example/ck_tile/17_grouped_gemm/quant_grouped_gemm_bf8_bquant.cpp
@@ -1,0 +1,7 @@
+// Copyright (c) Advanced Micro Devices, Inc., or its affiliates.
+// SPDX-License-Identifier: MIT
+
+#include "quant_run_grouped_gemm_example.hpp"
+
+template int run_gemm_example_persistency<ck_tile::bf8_t, ck_tile::QuantType::BQuantGrouped>(
+    const ck_tile::ArgParser&, std::string, std::string, bool);
